requirements.txt

torch>=2.0.0
torchvision>=0.15.0
numpy>=1.21.0
Pillow>=9.0.0

# Web framework
fastapi>=0.100.0
uvicorn[standard]>=0.22.0
python-multipart>=0.0.6

# Frontend
gradio>=4.0.0

# Utilities
requests>=2.28.0
pydantic>=2.0.0
python-dotenv>=1.0.0

# Development
pytest>=7.0.0
pytest-asyncio>=0.21.0